+++ ReleaseNotes.md
@@ -1,6 +1,9 @@
-# 2020.3.11.19xxx-pre-release
+## Never mind the fluff -- How do I use this?
 
-Proof of concept
-* `JustifySuppressionRule` (`Dixon.Design#DX0001`), the "Hello, World!" of FxCop rules.  A port of the well travelled example rule to ensure that `SuppressMessage` attributes include a justification.  See e.g. http://www.binarycoder.net/fxcop/html/ex_specifysuppressmessagejustification.html
+[Here's the recipe I use](https://github.com/SteveGilham/altcode.dixon/wiki).
+
+# 2020.xx.xx.xxxxx-pre-release
+
+* `ReraiseCorrectlyRule` (`Dixon.Design#DX0002`) : only `throw` and exception you've just created
 
 For previous releases, go here -- https://github.com/SteveGilham/altcode.dixon/blob/master/ReleaseNotes%20-%20Previously.md